// src/interfaces/overtime.interface.ts
export interface OvertimeCreate {
  //   employeeName: string;
  //   timeIn: string;
  //   timeOut: string | any;
  // empId: string;
  // empName: string;
  date: any;
  overtimeHours: string | any;
  extraOvertimeHours: string | any;
  overtimeAmount: string | any;
  hourlyAmount: string | any;
}

export interface OvertimeAttandance {
  overtimeHours: string | any;
}